// src/app/data/interfaces/IProducto.ts
export interface IProducto {
  codProducto?:         number;
  descripcionProducto?: string;
  estado?:              boolean;
  observacion?:         string;
  precioCompra?:        number;
  precioVenta?:         number;
  stockAlmacen?:        number;
  stockCompra?:         number;
  codCategoria?:        number;
}